pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.65.0",
    "claude-code-sdk>=0.0.20",
    "fastapi>=0.116.1",
    "google-cloud-bigquery>=3.36.0",
    "google-cloud-run>=0.11.0",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
    "websocket-client>=1.8.0",
    "websockets>=15.0.1",
]
